Move json utilities to own files and write tests for current state
