+++ src/utilities/index.ts
@@ -15,11 +15,13 @@ export * from './math'
 export * from './numbers'
 export * from './object'
 export * from './parameters'
+export * from './parseUnknownJson'
 export * from './processTypes'
 export * from './routes'
 export * from './savedFilters'
 export * from './seconds'
 export * from './state'
+export * from './stringifyUnknownJson'
 export * from './strings'
 export * from './subscriptions'
 export * from './timeIntervals'
