+++ src/utilities/json.ts
@@ -1,4 +1,4 @@
-import { isRecord, isString } from '@/utilities'
+import { isRecord } from '@/utilities'
 
 type JsonStringify = Parameters<typeof JSON.stringify>
 type JsonValue = JsonStringify[0]
@@ -6,37 +6,6 @@ type JsonReplacer = JsonStringify[1]
 
 export function stringify(value: JsonValue, replacer?: JsonReplacer): string {
   return JSON.stringify(value, replacer, 2)
-}
-
-export function parseUnknownJson(value: unknown): unknown {
-  // If the incoming value is a string, we attempt to parse it as JSON.
-  if (isString(value)) {
-    try {
-      const parsed = JSON.parse(value)
-
-      // If the parsed value isn't a string, we return it.
-      if (!isString(parsed)) {
-        return parsed
-      }
-
-      // Otherwise, we return the original value, since strings are valid JSON.
-      return value
-    } catch {
-      // silence is golden
-    }
-  }
-
-  return value
-}
-
-export function stringifyUnknownJson(value: unknown): string | null | undefined {
-  const parsed = parseUnknownJson(value)
-
-  if (isString(parsed)) {
-    return parsed
-  }
-
-  return JSON.stringify(parsed)
 }
 
 export function isValidJsonRecord(value: unknown): value is string {
